refactor: 패키지명 변경 및 record 적용
